[refactor] refactor HomePage to use RestaurantCard component
