+++ client/src/App.tsx
@@ -1,5 +1,5 @@
 // import React from "react";
-import HomePage from "./components/HomePage";
+import HomePage from "./pages/HomePage";
 
 const App = () => {
   return <HomePage />;
