+++ client/src/__tests__/HomePage.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, waitFor, within} from "@testing-library/react";
-import HomePage from "../components/HomePage";
+import HomePage from "../pages/HomePage";
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
 
 // mock data
