+++ client/vitest.config.ts
@@ -6,6 +6,6 @@ export default defineConfig({
     test: {
         globals: true,
         environment: 'jsdom',
-        setupFiles: './src/setupTests.ts',
+        setupFiles: './src/__tests__/setupTests.ts',
     },
 });
